# perf(lsp): load buffer contents once when processing semantic tokens responses